+++ R/DfExtractDataset.R
@@ -5,7 +5,8 @@
 #' @usage DfExtractDataset(dataset, trts, rdrs) 
 #' 
 #' 
-#' @param dataset The original dataset from which the subset is to be extracted
+#' @param dataset The original dataset from which the subset is to be extracted;
+#'    can be ROC, FROC or LROC
 #' @param trts A vector contains the indices of the modalities to be extracted. 
 #'    \strong{If this parameter is not supplied, all modalities are extracted.}
 #' @param rdrs A vector contains the indices of the readers to be extracted. 
@@ -20,10 +21,11 @@
 #' 
 #' @examples 
 #' ## Extract the data corresponding to the second reader in the 
-#' ## first modality from an include ROC dataset
+#' ## first modality from an included ROC dataset
 #' dataset1_2 <- DfExtractDataset(dataset05, trts = 1, rdrs = 2)
 #' 
-#' ## Extract the data of the first and third reader in all modality from the include ROC dataset
+#' ## Extract the data of the first and third reader in all 
+#' ## modality from the included ROC dataset
 #' datasetA_123 <- DfExtractDataset(dataset05, rdrs = c(1, 3))
 #' 
 #' @export
@@ -51,17 +53,36 @@ DfExtractDataset <- function(dataset, trts, rdrs){
     rdrs <- 1:J
   }
   
-  K <- dim(dataset$NL)[3]
-  K2 <- dim(dataset$LL)[3]
-  maxNL <- dim(dataset$NL)[4]
-  maxLL <- dim(dataset$LL)[4]
-  NL <- dataset$NL[trts, rdrs, , ]
-  dim(NL) <- c(I, J, K, maxNL)
-  dataset$NL <- NL
-  LL <- dataset$LL[trts, rdrs, , ]
-  dim(LL) <- c(I, J, K2, maxLL)
-  dataset$LL <- LL
-  dataset$modalityID <- dataset$modalityID[trts]
-  dataset$readerID <- dataset$readerID[rdrs]
-  return(dataset)
+  if (dataset$dataType != "LROC") {  
+    K <- dim(dataset$NL)[3]
+    K2 <- dim(dataset$LL)[3]
+    maxNL <- dim(dataset$NL)[4]
+    maxLL <- dim(dataset$LL)[4]
+    NL <- dataset$NL[trts, rdrs, , ]
+    dim(NL) <- c(I, J, K, maxNL)
+    dataset$NL <- NL
+    LL <- dataset$LL[trts, rdrs, , ]
+    dim(LL) <- c(I, J, K2, maxLL)
+    dataset$LL <- LL
+    dataset$modalityID <- dataset$modalityID[trts]
+    dataset$readerID <- dataset$readerID[rdrs]
+    return(dataset)
+  } else {
+    K <- dim(dataset$NL)[3]
+    K2 <- dim(dataset$LLCl)[3]
+    maxNL <- 1
+    maxLL <- 1
+    NL <- dataset$NL[trts, rdrs, , ]
+    dim(NL) <- c(I, J, K, maxNL)
+    dataset$NL <- NL
+    LLCl <- dataset$LLCl[trts, rdrs, , ]
+    dim(LLCl) <- c(I, J, K2, maxLL)
+    LLIl <- dataset$LLIl[trts, rdrs, , ]
+    dim(LLIl) <- c(I, J, K2, maxLL)
+    dataset$LLCl <- LLCl
+    dataset$LLIl <- LLIl
+    dataset$modalityID <- dataset$modalityID[trts]
+    dataset$readerID <- dataset$readerID[rdrs]
+    return(dataset)
+  }
 }