+++ R/SsSampleSizeKGivenJ.R
@@ -27,7 +27,7 @@
 #'
 #'
 #'
-#'@note The procedure is valid for ROC studies only; for FROC studies see \link{SsFROCPowerGivenJK}.
+#'@note The procedure is valid for ROC studies only; for FROC studies see Online Appendix Chapter 19.
 #'
 #'
 #'